refactor: move registration form from screens to components directory
